requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
flask>=2.0.0

# AI/ML frameworks
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.267
faiss-cpu>=1.7.4
gymnasium>=0.28.1
pettingzoo>=1.22.0
# ray[rllib]>=2.5.0  # Commented out due to installation issues

# Cybersecurity tools
# python-suricata>=0.1.0  # Commented out due to no matching distribution found
elasticsearch>=8.8.0
elasticsearch-dsl>=8.8.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0
streamlit>=1.24.0
websockets>=11.0.3

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
 
 Flask>=2.0.0
flask-cors>=5.0.1
flask-limiter>=3.5.0
numpy>=1.24.0
torch>=2.0.0
python-dotenv>=1.0.0